+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Upcoming
 
+- `apollo-language-server`
+  - Fix edge case for empty operations [#959](https://github.com/apollographql/apollo-tooling/pull/959)
+
 ## `apollo@2.4.1`
 
 - `apollo` 2.4.1